- move toast vuex to component register - fix storybook webpack error
