+++ components/singleton/toast.vue
@@ -8,6 +8,63 @@
 </template>
 
 <script>
+import Promise from 'bluebird'
+
+// keep it out of store state,
+// as it dont need to be reactive
+// and bluebird promise will update it self which cause vuex warning
+let promises = []
+
+let state = {
+  msgs: [],
+}
+
+
+const getters = {
+  currentMsg(state) { return state.msgs.length? state.msgs[0]: null },
+  length(state) { return state.msgs.length },
+}
+
+const mutations = {
+  ADD(state, payload) {
+    let resolve, reject
+    let promise = new Promise((res, rej) => {
+      resolve = res
+      reject = rej
+    })
+    promises.push({promise, reject, resolve})
+
+    payload.id = Date.now()
+    state.msgs.push(payload)
+  },
+
+  REMOVE(state) {
+    state.msgs.shift()
+    promises.shift()
+  },
+
+  RESOLVE(state, payload) {
+    if (promises.length === 0) return console.warn('No Toast to call action')
+    promises[0].resolve(payload)
+  },
+}
+
+const actions = {
+  add({ commit, getters, }, payload) {
+    if (typeof(payload) === 'object') {
+      commit('ADD', {
+        text: payload.text,
+        action: payload.action,
+      })
+    } else if (typeof(payload) === 'string') {
+      commit('ADD', { text: payload, })
+    }
+    return promises[getters.length - 1].promise
+  },
+
+}
+
+const namespaced = true
 
 export default {
   data() {
@@ -47,6 +104,16 @@ export default {
     },
 
   },
+  beforeCreate() {
+    if (this.$store) {
+      this.$store.registerModule('toasts', { namespaced, state, getters, mutations, actions })
+    } else {
+      console.error('Toast components need vuex to work');
+    }
+  },
+  destroyed() {
+     this.$store.unregisterModule('toasts')
+  }
 }
 </script>
 
@@ -124,7 +191,7 @@ button {
           </div>
 
           <div class="field">
-            <button class="button" type="button" @click="alert()">Notify and alert</button>
+            <button class="button" type="button" @click="alert()">Notify and Log the time</button>
           </div>
         </form>
       </div>
